Log file fix
Changed dictionary to use "HitPatternID" like the header declaration does, the error caused by this mistake is insignificant, the data was simply stored in a different data column instead and the correct one was left empty.
Data analysis will use the incorrect name, if any future tests were to be done then the correct name would be used for data analysis instead.
If comparison of old and new data is desired, the old CSV files can simply be edited to remove the empty-but-correctly-named column, then rename the full-but-incorrectly-named column to use to correct name.