Updated data tracking
Been messing around a bit with the data tracking do include information that would have been useful to have from the tests.
Game now tracks when the player starts and ends an obstacle pattern
Game also tracks when an intermission is happening

These pieces of information would have made it easier to isolate notable instances of tracking information for comparison purposes, as without them there is no way to determine exactly when a player is engaging with a pattern unless they get hit, but even then that only tells us that the pattern was active at that moment in time.