Mapping responsiveness fix
Remembered that some testers had called the mapping unresponsive and tested to find out why
Realized the threshold only considered the difference based on movement over a number of frames and had no logic for difference to the last handRotation that set the position, effectively allowing really slow rotations to fail at updating the player's position, even though the handRotation had changed significantly.
The mapping conditions now account for the last handRotation that positioned the avatar, and will move the avatar if the rotation difference is higher than that threshold in addition as an alternate condition to the existing one.